docs: covert to docs snippets2
