+++ apps/docs/src/guide/cookbook/combining-utxos.md
@@ -8,4 +8,4 @@ One way to avoid these errors is to combine your UTXOs. This can be done by perf
 
 > **Note:** You will not be able to have a single UTXO for the base asset after combining, as one output will be for the transfer, and you will have another for the fees.
 
-<<< @/../../docs-snippets/src/guide/cookbook/combining-utxos.test.ts#combining-utxos{ts:line-numbers}
+<<< @/../../docs-snippets2/src/guide/cookbook/combining-utxos.test.ts#combining-utxos{ts:line-numbers}
